Only add annoying pink banner to httpdmulti sites
There are sites running on merope and sterope that aren't staging sites
meant for viewing by clients (e.g., the WDT Handbook and GeoServer).
Displaying the banner on those sites only annoys *us*.

As httpdmulti is used *only* for running staging sites, this is an easy
(albeit simplistic) way to apply the banner where we want it.